+++ TractorMod/Framework/Attachments/HoeAttachment.cs
@@ -67,6 +67,13 @@ namespace Pathoschild.Stardew.TractorMod.Framework.Attachments
             if (this.Config.DigArtifactSpots && tileObj?.ParentSheetIndex == HoeAttachment.ArtifactSpotItemID)
                 return this.UseToolOnTile(tool, tile, player, location);
 
+            // harvest ginger
+            if (this.Config.HarvestGinger && tileFeature is HoeDirt dirt && dirt.crop?.whichForageCrop.Value == Crop.forageCrop_ginger && dirt.crop.hitWithHoe((int)tile.X, (int)tile.Y, location, dirt))
+            {
+                dirt.destroyCrop(tile, showAnimation: false, location);
+                return true;
+            }
+
             // till plain dirt
             if (this.Config.TillDirt && tileFeature == null && tileObj == null && this.TryStartCooldown(tile.ToString(), this.TillDirtDelay))
                 return this.UseToolOnTile(tool, tile, player, location);