+++ DataLayers/release-notes.md
@@ -1,4 +1,8 @@
 # Release notes
+## Upcoming release
+* Updated for Stardew Valley 1.5, including...
+  * new fertilizer types.
+
 ## 1.13
 Released 04 November 2020.